backend/requirements.txt
Flask==0.12.2
Werkzeug<0.13.0,>=0.12.0
google-api-python-client
google-cloud-language==1.0.0
google-cloud
gunicorn==19.7.1
